#1701: extract class-of-service types into types_cos.go
Moves the CoS type definitions verbatim out of types.go into
types_cos.go (same package config):

  ClassOfServiceConfig, CoSForwardingClass, CoSDSCPClassifier(+Entry),
  CoSIEEE8021Classifier(+Entry), CoSDSCPRewriteRule(+Entry),
  CoSScheduler, CoSSchedulerMap(+Entry), CoSInterface, CoSInterfaceUnit,
  CoSFairnessExpectation.

Note the deliberate split from the unrelated time-range SchedulerConfig
([edit schedulers], used by Policy.SchedulerName) which is a security
policy construct and lands in types_security.go, not here — the queue
scheduler CoSScheduler stays in CoS. Pure byte-identical motion;
consumers untouched. Verified: go build ./... clean, vet clean, tests pass.
